Move integer functions to separate module
